HomeDBManaager: remove doc access for disallowed users
If there is an existing websocket connexion, disallowed users might
still be able to use it to send data. There is a 5 second cache on the
doc implementation, but permissions are rechecked after the cache
expires. Checking if a user is disabled here when getting the document
should prevent the user from doing any further work on a document.